mcts.hpp: VisitStats::value_sum should be float, not double
Every other numeric value in this codebase (evaluate(), hp_fraction,
sample_damage_fraction, exploration_constant, ...) is float - double
here was an unjustified inconsistency with no real precision need for
this use case, caught while working out select_ucb1_action's own score
type.

+++ cpp/include/be/mcts.hpp
@@ -37,7 +37,7 @@ namespace be {
 // a given node's actual legal set).
 struct VisitStats {
   int visits = 0;
-  double value_sum = 0.0;
+  float value_sum = 0.0f;
 };
 
 // Sentinel used in a packed child key (see pack_action_pair below) for "this
